requirements_minimal.txt
# Minimal requirements for fast Vercel deployment
fastapi==0.116.1
pydantic==2.11.7
requests==2.32.4
google-generativeai==0.8.5
pdfplumber==0.11.7
